d use double

+++ json/test.d
@@ -6,9 +6,9 @@ int main(string[] args) {
   string text = readText("./1.json");
   JSONValue doc = parseJSON(text).object;
   JSONValue[] coordinates = doc["coordinates"].array;
-  float x = 0;
-  float y = 0;
-  float z = 0;
+  double x = 0;
+  double y = 0;
+  double z = 0;
 
   for (int i = 0; i < coordinates.length; i++) {
     JSONValue coord = coordinates[i];
